refactor: 'GeneArgs' struct in 'args.rs'
Updated 'GeneArgs' structure in 'args.rs' to accept multiple package names for action and added an install arg. Previously, the structure only allowed a single package name to be inputted. This change will improve flexibility and the overall functionality of the gene-utils program. A typo in a comment has also been corrected.