docs: add initial README for newly introduced 'gene-pmget' package
This commit adds a README file for the 'gene-pmget' module, a new feature currently under development. The module is designed to enhance the use of the GENE Cross Platform package installation system by providing support for various registries through the GPacistry. This is a crucial step toward the implementation of this feature.