refactor: revise gene-proc derive macro and enhance functionality
This commit refactors the derive macro in the gene-proc library. Renamed the function to 'derive_name' and made it more versatile to handle not only lifetimes but also non-lifetime variables. Made changes to return lower cased name of the struct as a Static Str. Now this macro will allow users to derive a 'name' function that returns the name of the struct in lower case.