refactor: update visibility of mod `raw_args`
Changed `raw_args` module visibility to public in the gene-pmm library. This update is necessary to allow other modules in the library to access and use the functionality of `raw_args`. The accessibility of raw arguments will be useful for detecting user inputs and controlling the behavior of the add function.