docs: update project report
The report's structure and content have been significantly updated. Notable changes include a thorough introduction section, in-depth explanations about the development and purpose of the GENE package manager, and a list of commands from various package managers. Coding principles and project aims are explained to provide a clearer understanding of the project. Moreover, information about the GPacistry and other key aspects of the package manager has been added. This offers a broader overview of the project and greater clarity its objectives.