docs: update README.md with instructions for adding new package manager
The documentation was expanded to include detailed instructions on how to add new package managers to the project. This change was made to provide clear guidance for future contributors on how to make additions to the project's functionality. Instructions include creating a rust module and adding the new package manager to the backends configuration.