docs: add example report from one of my prior projects
A detailed project report comes as an addition to the docs. It provides a walkthrough of the innovative project "ENKARM", a therapeutic robot designed for patients with limb disorders. Detailing from project synopsis through methods of design, to findings and suggestions, this report will serve as informative material to anyone interested in similar fields of technology and health. It further highlights the user-friendly, cost-effective, and accessible features of the robot, making the project unique and noteworthy. The documentation also outlines future enhancements that can broaden the application in other disorders, making it a valuable resource for researchers and developers.